os-net-config with additions to support static IP assignment
This version of os-net-config adds a Subnet object with logic for
assigning IP addresses out of a definable host IP range. The
subnets are defined in the os-net-config configuration file, with
subnets under the subnets: element in yaml. Each subnet should
have a name, an IP network/netmask, and a host_ip_range. Using
tokens in the network_config portion of the yaml file allows for
replacement with the appropriate data from the subnet object (IP
address, gateway, etc.). Unit tests are provided for the Subnet
object and the token replacement logic in os_net_config.utils.